refactor : dto 분리 및 테스트 코드 적용
